automated/linux/ptest: Analyze each test in package tests
Currently ptest.py analyze only exit code of each package test
to decide if it passed or not. However, ptest-runner can return
success code even though some tests failed. So we need to parse
test output and analyze it.

It also quite useful to see exactly which tests failed. So results are
recorded for each particular test, and lava-test-set feature is used
to distinguish packages.

Signed-off-by: Oleksandr Terentiev <oterenti@cisco.com>